Firewall setup for LTE modem link
This sets up a firewall which protects Aero board to be exposed
to Internet. Only access via ssh/port 22 is enabled.
However routing and NAT is set up the way that clients conencted
to WiFi can reach Internet also and NAT takes care that answers
are directed to the calling party.
This firewall is using iptables/netfilter tables filter and nat.

Signed-off-by: Johannes Fischer <johannes.fischer@intel.com>